Point 3, Modified Point 4, Add missing link
Distributed Ledger Technology (DLT) > Blockchains (Continued)

3. Definition of block shouldn’t be split up with mention of bitcoin
concepts in between. I put the bitcoin example after block is
explained, for better concept flow.

4. Nonce relates to the Bitcoin example, so indicate that this list of 4
pieces of metadata refer to the Bitcoin example (which continues with
the Merkle Tree section in the next slide).

Bonus: "or a brief history of blockchain technology, please click here."
was missing the link "here". Added it.

Signed-off-by: nathalie-ckc <nathalie.ckc.123@gmail.com>